+++ CHANGELOG.md
@@ -5,7 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
---
+- Converted to use Commandment Dispatcher as Controller
+- Removed Archetype Normalizer
+- Removed Task Hook interfaces
+- Ported HasTask to Commandment
 
 ---